# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Interview Agent RAG API
  version: 1.0.0
  description: |
    面试知识库 RAG 系统后端 HTTP 接口（主要端点）。

    链路：文档上传（两段式 prepare/commit）→ 切片 + 双向量嵌入（text_vector + question_vector）
    → 混合检索（语义 + 问题锚点 + 2-gram 关键词加权，低分级联 HyDE）→ RAG 流式问答（AI SDK data-stream）。

    ## 认证
    - 默认 `AUTH_MODE=disabled`（本地单机）：所有 /api 数据请求视为 local 单一用户，无需令牌。
    - 设置 `AUTH_TOKENS`（user-token 模式）后：除 /api/health、/api/metrics、/api/management 外的
      /api 路由必须携带 `Authorization: Bearer <用户令牌>`，跨用户数据互相不可见。
    - /api/management/* 始终要求 `Authorization: Bearer <ADMIN_TOKEN>`。

    ## 流式端点
    /api/knowledge/ask 与 /api/chat 返回 AI SDK data-stream 协议（Content-Type: text/plain; charset=utf-8），
    按行分片：`0:"文本分片"`（增量正文）、`2:[{...}]`（message annotations，如检索命中卡片）、
    `d:{...}`（结束帧，含 finishReason/usage）。客户端应逐行解析。
servers:
  - url: http://localhost:3000
    description: 本地开发 / Docker 部署
...
paths:
...
  /api/management/overview:
    get:
      tags: [management]
      summary: 运行总览（工具/工作流启停状态 + 运行统计）
      responses:
        '200': { description: 管理总览 }
        '401': { description: 管理员令牌缺失或无效 }
...
  /api/management/models/settings:
    put:
      tags: [management]
      summary: 更新模型运行时设置（qwen3 思考模式开关），热生效
      description: |
        thinking=false 时对 qwen3 系列模型的 chat 请求注入 reasoning_effort:'none'
        （OpenAI 标准参数，Ollama 兼容端点支持），跳过思维链直接作答。
        仅影响 qwen3* 模型；随 models.json 持久化。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                thinking: { type: boolean, description: 思考模式（默认 true 开启） }
      responses:
        '200': { description: 更新后的设置 }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        管理接口用 ADMIN_TOKEN；user-token 模式下数据接口用用户令牌。
        AUTH_MODE=disabled（默认本地模式）时无需认证。
...
security:
  - bearerAuth: []